Restructure project, migrate to scikit-build-core
- Move the Python code into the 'src' layout. Separate the C++ sources and headers into different directories. Update the CMakeLists.txt file to match the new folder structure and include the paths.
- Migrate from `setup.py` to `pyproject.toml`
- Replace the setup for `setuptools`, `CMake` and `PyBind11` with the setup for `scikit-build-core`. This will remove support for versions below Python 3.8 and substantially reduce the installation size by removing the inclusion of `PyBind11` in the package. It also adds support for modern versions, such as Python 3.11.
- Replace `std::bind2nd` with `std::bind`, as `std::bind2nd` was deprecated in C++11 and removed in C++17.
- Update github test workflow
