+++ cpp/datastructure/DataStructure.cpp
@@ -10,7 +10,7 @@
 #include <cmath>
 #include <functional>
 
-#include "DataStructure.h"
+#include "sacf/DataStructure.h"
 
 
 const char* EmptyDataStructureException::what() const throw(){
