+++ cpp/datastructure/datastructure_bindings.cpp
@@ -1,4 +1,4 @@
-#include "DataStructure.h"
+#include "sacf/DataStructure.h"
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
