+++ tox.ini
@@ -4,15 +4,16 @@
 # and then run "tox" from this directory.
 
 [tox]
-envlist = py27, py36, py37, py38, py39
+envlist = py38, py39, py310, py311, py312, py313
 
 [gh-actions]
 python = 
-  2.7: py27
-  3.6: py36
-  3.7: py37
   3.8: py38
   3.9: py39
+  3.10: py310
+  3.11: py311
+  3.12: py312
+  3.13: py313
 
 [testenv]
 deps =
